pyproject.toml
[project]
name = "langchain-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit",
    "langchain",
    "langgraph>=0.6.7",
    "langchain-google-genai",
    "google-generativeai",
    "python-dotenv>=1.1.1",
    "pydantic"
]